requirements.txt
torch>=2.2.0
transformers>=4.44.0
accelerate>=0.33.0
datasets>=2.20.0
peft>=0.11.0
trl>=0.9.0
bitsandbytes>=0.43.0
deepspeed>=0.15.0
vllm>=0.5.5
pyyaml>=6.0.0
tqdm
numpy
